Renaming packages and files to limit what needs to be replaced when using the template.
