+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.adammcneilly.androidapptemplate"
+        applicationId "com.adammcneilly.TODOREPLACE"
         minSdk 21
         targetSdk 31
         versionCode 1
