+++ app/src/androidTest/java/com/template/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.adammcneilly.androidapptemplate
+package com.template
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -17,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.adammcneilly.androidapptemplate", appContext.packageName)
+        assertEquals("com.template", appContext.packageName)
     }
 }
